requirements.txt
# 图像处理库
Pillow>=10.0.0

# 拖拽功能支持
tkinterdnd2>=0.3.0

# 构建工具依赖
pyinstaller>=6.0.0